fix signed/unsigned comparison warnings

+++ TFMatcher/src/AlignedTF.cpp
@@ -68,7 +68,7 @@ double AlignedTF::Distance()
 	double currentX, nextX;
 	double valueA, valueB;
 
-	for (int i = 0; i < sharedRangeEvents.size() - 1; i++)
+	for (size_t i = 0; i < sharedRangeEvents.size() - 1; i++)
 	{
 		currentX = sharedRangeEvents[i];
 		valueA = turningFunctionA.ValueAt(currentX);
